packages/py/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trustproof"
version = "0.2.0"
description = "Signed, verifiable action receipts for humans + AI agents. Ed25519 JWT with tamper-evident chain."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [{ name = "Ataraxia Labs", email = "protocol@ataraxialab.ai" }]
keywords = [
    "trust", "verification", "ai-agents", "jwt", "ed25519",
    "audit-trail", "cryptography", "ai-safety", "langchain",
    "openclaw", "verifiable-credentials", "fintech"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security :: Cryptography",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "PyJWT>=2.8.0",
    "cryptography>=41.0",
]

[project.urls]
Homepage = "https://ataraxia-labs-ai.github.io/TrustProof-Protocol/"
Repository = "https://github.com/ataraxia-labs-ai/TrustProof-Protocol"
Documentation = "https://ataraxia-labs-ai.github.io/TrustProof-Protocol/"
Changelog = "https://github.com/ataraxia-labs-ai/TrustProof-Protocol/blob/main/CHANGELOG.md"
Issues = "https://github.com/ataraxia-labs-ai/TrustProof-Protocol/issues"

[project.scripts]
trustproof = "trustproof.__main__:main"

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov", "mypy", "ruff"]

[tool.hatch.build.targets.wheel]
packages = ["src/trustproof"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
